@@ -1,14 +1,14 @@
 plugins {
 	id 'java'
-	id 'org.springframework.boot' version '2.7.17-SNAPSHOT'
-	id 'io.spring.dependency-management' version '1.0.15.RELEASE'
+	id 'org.springframework.boot' version '3.1.5'
+	id 'io.spring.dependency-management' version '1.1.3'
 }
 
 group = 'com.personalData'
 version = '0.0.1-SNAPSHOT'
 
 java {
-	sourceCompatibility = '11'
+	sourceCompatibility = '17'
 }
 
 configurations {
@@ -24,8 +24,8 @@ repositories {
 }
 
 ext {
-	set('springBootAdminVersion', "2.3.0")//"3.1.5") //"2.3.0") //añadido
-	set('springCloudVersion', "2021.0.8")
+	set('springBootAdminVersion', "3.1.5")
+	set('springCloudVersion', "2022.0.4")
 }
 
 dependencies {
@@ -38,7 +38,7 @@ dependencies {
 	implementation 'org.springframework.cloud:spring-cloud-starter-config'	//
 
 	implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client' //
-	implementation 'org.springframework.cloud:spring-cloud-starter-sleuth'	//añadido
+	//implementation 'org.springframework.cloud:spring-cloud-starter-sleuth'	//añadido
 
 	compileOnly 'org.projectlombok:lombok'
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
