@@ -2,7 +2,7 @@ package com.personalData.PersonalData.PersonalDataService.entity;
 
 import lombok.Data;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.time.LocalDateTime;
 
 @Data
