@@ -15,7 +15,7 @@ import java.util.List;
 @CrossOrigin
 @RestController
 @RequestMapping("/api/statusContract")
-@Api(tags="Status Contract", value="Web Service RESTful of Status Contracts")
+//@Api(tags="Status Contract", value="Web Service RESTful of Status Contracts")
 public class StatusRest {
     private final StatusService statusContractService;
 
@@ -25,12 +25,6 @@ public class StatusRest {
 
     //Retornar todos los status
     @GetMapping(value = "/all", produces = MediaType.APPLICATION_JSON_VALUE)
-    @ApiOperation(value="List of Status Contract", notes="Method to list all status contracts")
-    @ApiResponses({
-            @ApiResponse(code=201, message="Status contracts found"),
-            @ApiResponse(code=404, message="Status contracts not found"),
-            @ApiResponse(code=501, message="Internal server error")
-    })
     public ResponseEntity<List<StatusContract>> findAllStatus() {
         try {
             List<StatusContract> statusContracts = statusContractService.getAll();
