@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface ContractService extends CrudService<Contract> {
 
-    public
-    List<Contract> findByDriverId(Long driverId) throws Exception;
-    List<Contract> findByClientId(Long clientId) throws Exception;
-    List<Contract> findByDriverIdAndClientId(Long driverId, Long clientId) throws Exception;
+//    public
+//    List<Contract> findByDriverId(Long driverId) throws Exception;
+//    List<Contract> findByClientId(Long clientId) throws Exception;
+//    List<Contract> findByDriverIdAndClientId(Long driverId, Long clientId) throws Exception;
 }
